Fix clippy warnings
The notable functional change here is that the "Public API" comment is
turned from a doc comment into a regular comment.  Previously it (along
with all the forward slashes) was displaying awkwardly in the docs.
Clippy was right in this case that the empty line indicated that it was
not intended to attach to the item below.